+++ data/Store.ts
@@ -15,6 +15,7 @@ import {
     FilterLike,
     FilterValueSource,
     parseFilter,
+    RecordId,
     StoreRecord,
     StoreRecordId,
     StoreRecordOrId,
@@ -267,7 +268,10 @@ export type StoreRecordIdSpec = string | ((data: PlainObject) => StoreRecordId);
  *
  * @mcpHint in-memory data store used by grids and other data components
  */
-export class Store
+export class Store<
+    TData extends PlainObject = PlainObject,
+    TId extends StoreRecordId = RecordId<TData>
+>
     extends HoistBase
     implements FilterBindTarget, FilterValueSource, GridFilterBindTarget
 {
@@ -316,7 +320,7 @@ export class Store
      * {@link loadData} or by loading a tree structure with `loadRootAsSummary` set to true.
      */
     @observable.ref
-    summaryRecords: StoreRecord[] = null;
+    summaryRecords: StoreRecord<TData, TId>[] = null;
 
     /** @internal - used internally by any StoreFilterField bound to this store. */
     @observable
@@ -441,7 +445,9 @@ export class Store
         }
 
         this.summaryRecords = rawSummaryData
-            ? castArray(rawSummaryData).map(it => this.createRecord(it, null, true))
+            ? (castArray(rawSummaryData).map(it =>
+                  this.createRecord(it, null, true)
+              ) as StoreRecord<TData, TId>[])
             : null;
 
         const records = this.createRecords(rawData, null);
@@ -595,7 +601,7 @@ export class Store
         }
 
         if (!isEmpty(summaryUpdateRecs)) {
-            this.summaryRecords = summaryUpdateRecs;
+            this.summaryRecords = summaryUpdateRecs as StoreRecord<TData, TId>[];
             changeLog.summaryRecords = this.summaryRecords;
         }
 
@@ -792,7 +798,7 @@ export class Store
         }
 
         if (!isEmpty(summaryUpdateRecs)) {
-            this.summaryRecords = summaryUpdateRecs;
+            this.summaryRecords = summaryUpdateRecs as StoreRecord<TData, TId>[];
             changeLog.summaryRecords = this.summaryRecords;
         }
 
@@ -859,37 +865,37 @@ export class Store
     }
 
     /** Records in this store, respecting any filter (if applied).*/
-    get records(): StoreRecord[] {
-        return this._filtered.list;
+    get records(): StoreRecord<TData, TId>[] {
+        return this._filtered.list as StoreRecord<TData, TId>[];
     }
 
     /** All records in this store, unfiltered.*/
-    get allRecords(): StoreRecord[] {
-        return this._current.list;
+    get allRecords(): StoreRecord<TData, TId>[] {
+        return this._current.list as StoreRecord<TData, TId>[];
     }
 
     /** All records that were originally loaded into this store.*/
-    get committedRecords(): StoreRecord[] {
-        return this._committed.list;
+    get committedRecords(): StoreRecord<TData, TId>[] {
+        return this._committed.list as StoreRecord<TData, TId>[];
     }
 
     /** Records added locally which have not been committed.*/
-    get addedRecords(): StoreRecord[] {
+    get addedRecords(): StoreRecord<TData, TId>[] {
         return this.allRecords.filter(it => it.isAdd);
     }
 
     /** Records removed locally which have not been committed.*/
-    get removedRecords(): StoreRecord[] {
+    get removedRecords(): StoreRecord<TData, TId>[] {
         return differenceBy(this.committedRecords, this.allRecords, 'id');
     }
 
     /** Records modified locally since they were last loaded. */
-    get dirtyRecords(): StoreRecord[] {
+    get dirtyRecords(): StoreRecord<TData, TId>[] {
         return this.allRecords.filter(it => it.isDirty);
     }
 
     /** Alias for {@link Store.dirtyRecords} */
-    get modifiedRecords(): StoreRecord[] {
+    get modifiedRecords(): StoreRecord<TData, TId>[] {
         return this.dirtyRecords;
     }
 
@@ -897,25 +903,24 @@ export class Store
      * Root records in this store, respecting any filter (if applied).
      * If this store is not hierarchical, this will be identical to 'records'.
      */
-    get rootRecords(): StoreRecord[] {
-        return this._filtered.rootList;
+    get rootRecords(): StoreRecord<TData, TId>[] {
+        return this._filtered.rootList as StoreRecord<TData, TId>[];
     }
 
     /**
      * Root records in this store, unfiltered.
      * If this store is not hierarchical, this will be identical to 'allRecords'.
      */
-    get allRootRecords(): StoreRecord[] {
-        return this._current.rootList;
+    get allRootRecords(): StoreRecord<TData, TId>[] {
+        return this._current.rootList as StoreRecord<TData, TId>[];
     }
 
     /**
      * Single summary data record, if only one (or null if none). Maintained for convenience and
      * for backwards compat with app code predating support for multiple {@link summaryRecords}.
      */
-    get summaryRecord(): StoreRecord {
+    get summaryRecord(): StoreRecord<TData, TId> {
         if (isNull(this.summaryRecords)) return null;
-
         throwIf(
             this.summaryRecords.length > 1,
             'Store has multiple summary records - must access via Store.summaryRecords.'
@@ -1078,13 +1083,12 @@ export class Store
      *      active filter is excluding it from the primary `records` collection. True to restrict
      *      matches to this Store's post-filter StoreRecord collection only.
      */
-    getById(id: StoreRecordId, respectFilter: boolean = false): StoreRecord {
+    getById(id: StoreRecordId, respectFilter: boolean = false): StoreRecord<TData, TId> {
         if (isNil(id)) return null;
         const summaryRecord = this.summaryRecords?.find(it => it.id === id);
         if (summaryRecord) return summaryRecord;
-
         const rs = respectFilter ? this._filtered : this._current;
-        return rs.getById(id);
+        return rs.getById(id) as StoreRecord<TData, TId>;
     }
 
     /**
@@ -1096,10 +1100,10 @@ export class Store
      * @param id - ID of record to be queried.
      * @param respectFilter - true to skip records excluded by any active filter.
      */
-    getChildrenById(id: StoreRecordId, respectFilter: boolean = false): StoreRecord[] {
+    getChildrenById(id: StoreRecordId, respectFilter: boolean = false): StoreRecord<TData, TId>[] {
         const rs = respectFilter ? this._filtered : this._current,
             ret = rs.childrenMap.get(id);
-        return ret ? ret : [];
+        return (ret ? ret : []) as StoreRecord<TData, TId>[];
     }
 
     /**
@@ -1111,10 +1115,10 @@ export class Store
      * @param id - ID of record to be queried.
      * @param respectFilter - true to skip records excluded by any active filter.
      */
-    getDescendantsById(id: StoreRecordId, respectFilter = false): StoreRecord[] {
+    getDescendantsById(id: StoreRecordId, respectFilter = false): StoreRecord<TData, TId>[] {
         const rs = respectFilter ? this._filtered : this._current,
             ret = rs.getDescendantsById(id);
-        return ret ? ret : [];
+        return (ret ? ret : []) as StoreRecord<TData, TId>[];
     }
 
     /**
@@ -1126,10 +1130,10 @@ export class Store
      * @param id - ID of record to be queried.
      * @param respectFilter - true to skip records excluded by any active filter.
      */
-    getAncestorsById(id: StoreRecordId, respectFilter: boolean = false): StoreRecord[] {
+    getAncestorsById(id: StoreRecordId, respectFilter: boolean = false): StoreRecord<TData, TId>[] {
         const rs = respectFilter ? this._filtered : this._current,
             ret = rs.getAncestorsById(id);
-        return ret ? ret : [];
+        return (ret ? ret : []) as StoreRecord<TData, TId>[];
     }
 
     /** True if the store is confirmed to be Valid. */
@@ -1395,7 +1399,7 @@ export class Store
             });
             ret.finalize();
             return ret;
-        });
+        }) as StoreRecord<TData, TId>[];
     }
 }